Add reset_payment_request button on Pfmp#show page + Display reason for unpaid + Copy changes + Mini refactor (#714)
Fixes #[692](https://github.com/betagouv/aplypro/issues/692) + #722 

NOTE: this contains the "partial" introduction of the `PfmpManager`
service. Its goal is to encapsulate complex tasks related to a `Pfmp`
record like calculating amount, rebalancing, launching new payment
requests, etc. I moved some of the specs that will eventually belong to
the service but that still rely, for now, on `after_save` callback. The
object itself is rather crude and a plain PORO for now.

- [x] Prevent payment reset at ~~API~~ Service level too 
- [x] Improve warning message for UI
- [x] Integrate :unpaid_reason under same accessor as rejection reason
in p_r decorator
- [x] Controller spec for new action
- [x] Change home button. -> Consulter et gérer les envois en paiment
- [x] Onglet: Envoyer en paiement -> "Paiements"
- [x] Factor in locales the cta message + warning message
- [x] E2E Scenario
    -> test list failed payments + click + reset payment
- [x] Potentially introduce a Pfmp Management service to factor this
feature + Rebalancing feature in one piece of BL?
-> ~~Move Amount Calculator concern to service~~ scope creep for this
PR, lets address this in a follow up~~
-> ~~Remove after_save callback for rebalancing amounts~~ scope creep
for this PR, lets address this in a follow up
- [x] Validate UX of button change
    -> ~~first propal:~~
    -> ~~second propal:~~
    -> ~~third propal~~
   -> 4th (validated):
![Screenshot 2024-04-16 at 14 52
16](https://github.com/betagouv/aplypro/assets/2844875/db7ccd56-cb18-4fa0-8b44-8df78d4c449c)