workers: add a dedicated worker for the payments
The payment tasks are quite heavy: sending requires validating 7000+
records with an XSD schema, updating the associated records, and
parsing the results alos means downloading big files which results in
big downloads + updating up to 100K records (and going through big XML
files again).

To avoid getting OOM-killed by Scalingo (which happens in a dedicated
container or in a worker), spawn a new worker to clearly handle that
side of the product, and we can beef up the specs when it's running in
Scalingo, at least for the time being where we deal with huge volumes.