jobs: don't retry the PollPaymentsServerJob for now
We're too early-stage to confidently and immediately retry this job
until it works. If it's failed once there is little chance it will
succeed a minute or an hour later.

We might reconsider in the future but just because this job is meant
to be run daily doesn't mean we want the whole
auto-retry-exponential-backoff baked in.