jobs: fix the PollPaymentsServerJob remove_file logic
There's a strange logic path in the ASP file removal, because we only
try to remove the file if we've successfully saved it on our side,
even if the processing goes wrong (ResponseFileParsingError): if the
file is safe with us there's no need to keep it on the ASP servers.

The problem is that `handler.file_saved?` was also triggering the
error, making this code crash:

      begin
        handler.parse!
      rescue ASP::Errors::ResponseFileParsingError, ASP::Errors::UnmatchedResponseFile => e
        Sentry.capture_exception(e)
      ensure
        ASP::Server.remove_file!(filename: filename) if handler.file_saved?
      end

In the case of an unmatched file (i.e we can't find the associated
ASP::Request, for payment files we're always good), the `parse!` code
will raise (persist_file! -> target_attachment-> record ->
find_record! -> UnmatchedResponseFile).

This goes to the rescue block where we tell Sentry about it, and then
the ensure block will try to remove the file if we've saved it. But
`file_saved?` is calling `target_attachment` too, which will trigger
the error again, but this time uncaught.

Solve this by simply making sure `file_saved?` always does it job by
gracefully replying no (which is the correct answer) if the file
wasn't even matched.