jobs: introduce the new payments automation architecture
This is how it goes:

1. ConsiderPaymentRequestsJob is given an `upto` date and will look at
all the pending payment requests for PFMPs that ended before that date
via the `ASP::PaymentRequest.to_consider` scope which does a
first (and very easy/fast) pass at filtering the "good" ones: this
means PFMFs with a RIB, an attributive decision and that's it. It will
enqueue a job for each of these payment requests, on the dedicated
payments worker.

2. PreparePaymentRequestJob takes a single payment request and runs
the more complicated machinery of `mark_ready!`: this runs all the
necessary checks and will even try and fetch the student's information
first to ensure we have fresh data.

3. SendPaymentRequestsJob remains the same for now and needs to be
called with a group of payment requests: in the future we might just
want to say "grab anything in the ready scope" but for now we're
keeping control.

This should yield a fair amount of resilience/performance since the main
performance bottleneck (checking/fetching data for a single request)
is contained in its own job that can be retried and queued alongside
hundreds/thousands of sibling jobs.