exclusion: add an exclusion model
This allows us to store and identify establishments which need to be
completely excluded (because for some reason the API doesn't send back
the right contract type), or because they have only certain specific
diplomas (i.e MEF codes) that are not covered.

It means those establihsments might be offer diplomas A & B but
diploma B is not under the right contract type so we have to store
these cases to properly discriminate against.