asp: replace soft hyphens in the rib name
OK this was really odd: yesterday we had a RIB name that was failing
the regexp check for mysterious reasons. Look at this:

"Marie­Curie"

Think that's a space in the middle? Think again, it's a soft
hyphen[1]. Its purpose is to maybe-break-lines-apart-if-necessary, at
least that's what I understood. Either way it's not caught by
`I18n.transliterate` so remove it before and tuck the character in its
own variable to avoid getting fooled again.

[1]: https://en.wikipedia.org/wiki/Soft_hyphen