asp: reduce the ASP::FileHandler to be ASP::FileSaver
This undoes the mess that the ASP::FileHandler was forced to do:
matching the 3 different types of file we get back from the ASP,
storing them on the right request/payment return then launching the
processing of the file.

Now it just saves to the right record (hopefully), which means we can
untangle the reading/processing functionality eslewhere and put it in
separate jobs.