db: add indices on ASP attributes
I don't understand why it wasn't done in the first place when we added
them but this will

1. ensure integrity because they actually all need to be unique in
their scope;

2. hopefully provide a solid performance boost especially in the
payments-parsing job which reads thousands of XML records and uses
`asp_prestation_dossier_id` to locate the matching record in our
database. Right now there is no index so it probably does a full-table
scans for each record which is bad. We're swapping hard and getting
OOMK on the latest file.

[1]: bc3ba1b4aacfed537a5f4556861ba5bbaf8ecaf5