jobs: remove the arguments to SendPaymentRequestsJob
We want the job to figure out for itself what payments we can send. We
had the parameter to ensure we kept control and could filter out what
was not ready, but now that all the filters have been backported in
the app we can trust the job to do the right thing.