jobs: don't use the capsule feature
Our sidekiq initializer says "use the single-threaded capsule for the
payments_serial queue" but that capsule is used for all workers, and
it doesn't seem to care about the -q argument passed to each worke
process.

Remove it in the meantime and just disable concurrency (with n=1) on
the payments worker, which kinda makes sense because it handles big,
unfrequent jobs that are all to do with sending or receiving files
from the ASP. We could potentially rename it to asp-worker to make
that clearer.

As part of that logic, move the PreparePaymentRequestJob to the normal
worker because like the other jobs there it does some API-querying and
operates on our app only.