sygne: fallback on the mef code if the national one is missing
We use `codeMefRatt` to figure out schoolings, which is the academic
MEF, see commit 427dc9f29d0812715ed0023e5bcf50d31d12f4c6.

In some cases we might not have it: the SYGNE `/scolarites` endpoint
does not provide `codeMefRatt` at the moment, see tracking issue:
https://github.com/betagouv/aplypro/issues/918.

In most scenarios the codeMefRatt is the same as the codeMef so
fallback on that if we don't have it.